move hls-parser
