+++ packages/hls-parser/resources/StateMachine.puml
