+++ packages/hls-parser/src/lib/stateMachine.ts
