+++ packages/hls-parser/src/lib/tags/base.ts
