+++ packages/hls-parser/src/lib/tags/tagWithAttributesProcessors.ts
