+++ packages/hls-parser/src/lib/types/parsedPlaylist.d.ts
