+++ packages/hls-parser/src/lib/types/sharedState.d.ts
@@ -1,4 +1,4 @@
-import type { Encryption, MediaInitializationSection, Segment, VariantStream } from '@/hls-parser/types/parsedPlaylist';
+import type { Encryption, MediaInitializationSection, Segment, VariantStream } from './parsedPlaylist';
 
 export interface SharedState {
   // Used to persist EXT_X_BITRATE across segments
