+++ packages/hls-parser/src/lib/utils/parse.ts
